тесты

+++ src/s21_string.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 
 #define s21_NULL (void *)0
-typedef unsigned long s21_size_t;
+typedef unsigned long long s21_size_t;
 
 void *s21_memchr(const void *str, int c, s21_size_t n);
 int s21_memcmp(const void *str1, const void *str2, s21_size_t n);
@@ -16,8 +16,6 @@ void *s21_memset(void *str, int c, s21_size_t n);
 // длина меньше) в конец dest
 // char *s21_strncat(char *dest, const char *src, s21_size_t n);
 
-// возвращает адрес символа c в строке s, начиная с головы, или NULL, если
-// строка s не содержит символ c
 char *s21_strchr(const char *str, int c);
 
 int s21_strncmp(const char *str1, const char *str2, s21_size_t n);
